refactor(visual-diff): split screenshot.ts → capture + index (SRP)
screenshot.ts (392 lines) had two responsibilities: per-element capture
and markdown index generation. Split into three files, each ≤300 lines:

- screenshot.ts (167) — orchestrator only; re-exports generateMarkdownIndex
- screenshot-capture.ts (173) — captureElement + per-overlay capture helpers
- screenshot-index.ts (88) — generateMarkdownIndex writer

Behavior-preserving: all exports (runScreenshots, generateMarkdownIndex,
ScreenshotOptions, ScreenshotResult) remain on screenshot.ts. Tests
unchanged (67/67 passing). No new typecheck errors.
